doc/meson: Remove install arg.
The project currently is set to meson version 0.49.0 but uses features
the install argument for configure_file was introduced with meson 0.50.0.

This patch removes the unused parameter fixes the following warnings:
```
WARNING: Project targeting '>=0.49.0' but tried to use feature introduced in '0.50.0': install arg in configure_file
WARNING: Project specifies a minimum meson_version '>=0.49.0' but uses features which were added in newer versions:
* 0.50.0: {'install arg in configure_file'}
WARNING: Deprecated features used:
* 0.50.0: {'install arg in configure_file'}
```

In an initial patch I just bumped the meson version, but this introduced
problems in our Debian CI.
An increase in the meson version would be possible by using a meson
distribution from pip, as it's done for old distros in xrdesktop.