aux/os/ble_dbus: Don't require power management characteristic to be notifiable.
At least the SteamVR 2.0 lighthouses have a power management characteristic
that does not have the "notify" flag, only "write". So before this change,
the command would not find any valid characteristics to write to and be
a no-op. The power management code path does not require it to be notifiable
in the first place, so this change just ignores that flag.

Part-of: <https://gitlab.freedesktop.org/monado/monado/-/merge_requests/2269>