st/oxr: return XR_SPACE_BOUNDS_UNAVAILABLE for unimplemented function
Fixes 730362a22

Since 730362a22 OXR_CHECK_XRET the return code
XRT_ERROR_COMPOSITOR_FUNCTION_NOT_IMPLEMENTED or (XRT_ERROR_NOT_IMPLEMENTED)
was translated to returning an XR_RUNTIME_FAILURE to the app with the
OXR_CHECK_XRET macro.

That the function is not implemented on an xrt_device is an implementation
detail in monado that we don't need to pass along to the application, the
application only needs to know that bounds are not available.

Part-of: <https://gitlab.freedesktop.org/monado/monado/-/merge_requests/2289>