u/pacing: Add U_PACING_COMP_TIME_FRACTION_PERCENT option
This adds the option U_PACING_COMP_TIME_FRACTION_PERCENT that allows
us to set the default compositor time as a percentage of the estimated
frame time.

This will be an easier way to tweak compositor time independently of
the particular HMD frame rate, when doing so is necessary for
performance reasons. This is currently often the case on NVIDIA
proprietary driver when used together with
XRT_COMPOSITOR_USE_PRESENT_WAIT=TRUE

Part-of: <https://gitlab.freedesktop.org/monado/monado/-/merge_requests/2526>